refactor: centralize shared types, separate lib from utils, colocate domain hooks
