--- src/components/profile/index.ts
@@ -1,1 +1,0 @@
-export * from "./personal-information-card";
