+++ src/services/event.ts
@@ -12,44 +12,11 @@ import {
   utcEndOfWeek,
   utcStartOfMonth,
   utcStartOfWeek,
-} from "@/lib/time";
+} from "@/utils/time";
 import type { CreateEvent, UpdateEvent } from "@/schema/event";
 import type { Event, EventType, RsvpStatus } from "@/generated/prisma/client";
-
-export interface EventWithRsvpCount extends Event {
-  rsvpCount: number;
-}
-
-export interface EventSummary {
-  id: number;
-  title: string;
-  startDate: Date;
-  endDate: Date;
-  isAllDay: boolean;
-  eventType: EventType;
-  location: string | null;
-  groupName: string | null;
-  rsvpCount: number;
-}
-
-export interface RsvpDetail {
-  memberId: number;
-  memberName: string;
-  status: RsvpStatus;
-  respondedAt: Date;
-}
-
-export interface RsvpCounts {
-  going: number;
-  maybe: number;
-  declined: number;
-}
-
-export interface InviteeDetail {
-  memberId: number;
-  memberName: string;
-  createdAt: Date;
-}
+import type { EventWithRsvpCount, EventSummary, RsvpDetail, RsvpCounts, InviteeDetail } from "@/types/event";
+export type { EventWithRsvpCount, EventSummary, RsvpDetail, RsvpCounts, InviteeDetail } from "@/types/event";
 
 export async function createEvent(data: Omit<CreateEvent, "memberIds" | "groupIds">, ownerid: number): Promise<Event> {
   try {
