--- src/services/index.ts
@@ -1,2 +1,0 @@
-export * from "./referral";
-export * from "./email";
