+++ src/types/member.ts
@@ -1,0 +1,17 @@
+export interface MemberSummary {
+  ownerid: number;
+  ownername: string;
+}
+
+export type GroupOption = {
+  id: number;
+  name: string;
+  memberCount: number;
+  memberIds: number[];
+};
+
+export type MemberOption = {
+  ownerid: number;
+  ownername: string;
+  photoUrl?: string | null;
+};
