+++ src/utils/photo-constraints.ts
